Limit database history size
Some users might have noticed that OP database grows a lot quite quickly
this is because OP is storing every event into the db so it can then show
it on the web page, problem is that there's no limit on how many lines
to store there. This patch adds a limit of 100 lines in history, leaving
only the last 50 lines each time the limit is reached.

modified:   django-web/openproximity/models.py